pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-codex-bridge"
version = "0.1.0"
description = "Standalone Discord to tmux Codex bridge"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "discord.py>=2.5,<3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[project.scripts]
discord-codex-bridge = "discord_codex_bridge.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
